third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-comparison-show-all.html

<html>
<head>
  <script src="../../inspector/inspector-test.js"></script>
  <script src="../resources/heap-snapshot-test.js"></script>
<script>

function test() {
  var instanceCount = 24;
  var firstId = 100;
  function createHeapSnapshotA() {
    return HeapSnapshotTestRunner.createHeapSnapshot(instanceCount, firstId);
  }
  function createHeapSnapshotB() {
    return HeapSnapshotTestRunner.createHeapSnapshot(instanceCount + 1, firstId + instanceCount * 2);
  }

  HeapSnapshotTestRunner.runHeapSnapshotTestSuite([function testShowAll(next) {
    HeapSnapshotTestRunner.takeAndOpenSnapshot(createHeapSnapshotA, createSnapshotB);
    function createSnapshotB() {
      HeapSnapshotTestRunner.takeAndOpenSnapshot(createHeapSnapshotB, step1);
    }

    function step1() {
      HeapSnapshotTestRunner.switchToView('Comparison', step2);
    }

    function step2() {
      var row = HeapSnapshotTestRunner.findRow('A');
      TestRunner.assertEquals(true, !!row, '"A" row');
      HeapSnapshotTestRunner.expandRow(row, step3);
    }

    var countA;
    var countB;
    function step3(row) {
      countA = row._addedCount;
      TestRunner.assertEquals(true, countA > 0, 'countA > 0');
      countB = row._removedCount;
      TestRunner.assertEquals(true, countB > 0, 'countB > 0');

      var buttonsNode = HeapSnapshotTestRunner.findButtonsNode(row);
      TestRunner.assertEquals(true, !!buttonsNode, 'buttons node (added)');
      var words = buttonsNode.showAll.textContent.split(' ');
      for (var i = 0; i < words.length; ++i) {
        var maybeNumber = parseInt(words[i], 10);
        if (!isNaN(maybeNumber))
          TestRunner.assertEquals(
              countA + countB - row._dataGrid.defaultPopulateCount(), maybeNumber, buttonsNode.showAll.textContent);
      }
      HeapSnapshotTestRunner.clickShowMoreButton('showAll', buttonsNode, step4);
    }

    function step4(row) {
      var rowsShown = HeapSnapshotTestRunner.countDataRows(row, function(node) {
        return node.data.addedCount;
      });
      TestRunner.assertEquals(countA, rowsShown, 'after showAll click 1');

      countB = row._removedCount;
      TestRunner.assertEquals(true, countB > 0, 'countB > 0');
      var buttonsNode = HeapSnapshotTestRunner.findButtonsNode(row);
      TestRunner.assertEquals(false, !!buttonsNode, 'buttons node (deleted)');

      var deletedRowsShown = HeapSnapshotTestRunner.countDataRows(row, function(node) {
        return node.data.removedCount;
      });
      TestRunner.assertEquals(countB, deletedRowsShown, 'deleted rows shown');
      setTimeout(next, 0);
    }
  }]);
}

</script>
</head>
<body onload="runTest()">
<p>
Tests Comparison view of detailed heap snapshots.
The &quot;Show All&quot; button must show all nodes.
</p>
</body>
</html>
